fetcher: improve network and no space left errors handling
Before this change, when a network error occured or if there is not
space left on the storage device, crawld would wait for one hour and
move on with the next repository. Now, it will wait for one hour and
retry the action it attempted, which makes much more sense.