errbag: add errbag package for throttling based on error rate
Package errbag allows the creation of an errbag. When the application
produces errors, it shall record them using Record() method which put
them into the errbag.
Once the errbag is full, a call to Record() will sleep for a predefined
amount of time before resuming. The errbag leaks errors at a predefined
rate as well so a certain amount error rate is allowed.

Of course, the errbag needs to be inflated for effective use and
deflated when of no use anymore.