repo: use errbag as throttler
It revealed to be impossible to rely on the type of errors returned by
the functions used by the fetcher to determine when the fetching task
shall pause (eg: when network is down or storage is full).

Hence, use errbag package to throttle on the error rate. When too many
errors are produced quickly, this is a sign that something real bad
happened, like the storage is full, permission is denied to write on the
clone directory, network is down or something of the like.

Use configuration options for the throttler to determine how much time
to wait before resuming the fetching task when too many errors arise,
what is a good sliding window size to determine the throttling or what
is the maximum tolerated error rate.