errbag: immediately kill the leak routine when Deflate() is called
With the previous implementation, it would the routine to empty the
error channel before noticing that it is closed and returning.
Depending on the leak rate and the number of errors in the channel, this
could take quite a while.

Refactor the code such that we add a new channel to signal to a parent
routine when we're done, which will exit on the signal. Killing the
parent obviously kills the child so we avoid the aforementioned problem.

While there, also rename airLeak to errLeak which shall have been its
name in the first place.