errbag: add callback function to Record() method
With a callback function, the caller is now able to know the status of
the throttler upon recording an error.

While here, simplify errLeak code and give it a stop signal through a
channel. If the stop signal is sent and errLeak() was in a wait state,
it will finish waiting before exciting. I think that this is acceptable.
The real deal would have been for errLeak() to empty the error channel
before returning.